Creates Makefile "help" target
What does the Makefile do? `make help` will tell you! Newcomers to the
project should be able to review the high-level commands available in
the developer environment without reading the Makefile directly. The
"help" target now provides this, and provides it in such a way that any
new targets will automatically be added to the help listing—provided
they use the "##" comment marker.