Creates "flake8" Makefile target
Runs flake8 to lint all Python files. Deferring to the `flake8` commands
on which files are Python, rather than explicitly passing a whitelist of
files to lint.

Users must install at least v3.3.0 of flake8; older versions, such as
2.5.5 (currently available in the Fedora 25 EOL repos), are not
sufficient for the file discovery logic. As a stopgap, we can pass "."
for the current working directory to flake8, which will force even older
versions to traverse for file discovery.

Note that we're *not* setting the `assert-dom0` target as a dependency,
since dom0 will not have flake8 installed by default. It's reasonable
for the dev env to have it, and we'll make sure the linting checks run
in CI.