Flake8 linting of base tests (E741)
Single commit for resolution of the E741 violation, which states that
"l" (i.e. lowercase L) should not be used as a variable name, since it
can be mistaken visually for a "1" (i.e. the number one). That's a bit
of a stretch in my mind, but hey, let's trust the linter.

The first version of flake8 including this check is v3.5.0. Older
versions will silently ignore the violation.