Creates platform tests for VM base OS
Partial progress toward #48.

Trying to write pass/fail tests for supported platforms for all VMs.
Aiming for a hard fail on EOL Fedora 25, in support of migration to Fedora
26, but focusing on the test structure for now.

The hardcoded copypasta in the form of the VM list is not DRY, and
suggests to me we should consider a more declarative format for the test
assumptions, e.g. a single YAML file describing the VM layout and
attributes, which all tests can import once and reuse.