Lint all Python files via flake8
For real this time! Added a `find` command to run `file` on every file
and execute flake8 on any that `file` identifies as a Python script.
Still running the `flake8 .` command separately. The disadvantage to
this approach is that there are two flake8 runs, so errors in the first
will cause the Makefile target to exit, without checking the other
files. That's acceptable: we'll still catch errors in CI and locally,
although it may take two runs of `make flake8` in order to catch them.