Adds tests for default DispVM template
We can't just test the running VMs for Fedora 26 compliance and call it
a day. DispVMs may still be set to Fedora 25, and we won't catch that
via interactive testing (i.e. using qvm-run to inspect the base OS).
Instead let's ask dom0 what the default template will be for the next
created DispVM, and confirm it's Fedora 26.