Wrapper script for build logic
Trying to encapsulate the command sequence used for managing the
grsec-patched TemplateVMs so others can easily reproduce the work.
For now the template repo is private, we'll sort that out soon.

Uses git logic from securedrop-workstation.conf in
qubes-template-securedrop-workstation to pull/validate source.

Failed builds will leave a lingering chroot that cannot be cleaned up
with a simple `rm -rf`. Let's use the bundled Makefile target for the
qubes-builder repo to tear down the chroot if it's likely to be there.