Add git tag signing key to keyring
The qubes builder uses a gpg keyring in qubes-builder/keyrings/git, in
order to verify the signature of the tag of the workstation template, we
must import our private key into this keyring and set it as trusted for
the build.

The key is currently set to my personal key, but it will be changed in
the near future to an FPF signing key.