Run flake8 tests for Python3 in container
Hacky implementation, I admit. As part of the migration from Python2 to
Python3, we must update the flake8 logic to use Python3, as well. The
most straightforward way to do so is in a container. In order to support
volume mounting in the container, we must update CircleCI to use the
"machine" type for builds.

The wrapper script logic is a bit convoluted here, would prefer to fold
in the 'file' package requirement in the underlying container
(freedomofpress/ci-python), but deferring that work for now to unblock
development on the workstation.

As for why the file package is necessary, flake8 autodiscovery will only
lint .py files, not the various scripts we have here.