Installs python-qt4 in whonix-ws-14
The `python-qt4` package is required for the GUI feedback integration,
so status messages can be displayed to the user. The package isn't
present by default, and we need it in sd-journalist. Since we don't
have SD TemplateVMs yet, we're adding the package to the base template,
installing in `whonix-ws-14`, rather than `sd-journalist`, to get
persistence for the package.