Forces use of gedit for text files via mailcap
This may be a larger problem with the Debian 9 transition: xdg-open
when run via a terminal (rather than invoked via e.g. Nautilus
integration, or otherwise in a GUI desktop environment) does not
honor the mimetype handlers declared in the desktop files. For
text/plain, edited the mailcap file in order to address.