Writes YAML config for securedrop-proxy
The YAML config file contains site-specific information such as the
Onion URL, and is required for the `securedrop-proxy` package to
function. Would be ideal to write the YAML intelligently, rather than as
line substitutions, this "just worked". In the future perhaps we can
port the Salt to Python and `import yaml`.

Includes tests. We know the Onion URL we want to connect to (for the
sd-whonix config), so let's read that from the JSON file on disk and
make sure the configuration landed on the target VM as intended.