Breaks out apt-test repo logic
Using a separate SLS file, since we need the apt-test repo configured in
more than just the `sd-workstation-template` VM: we need it at least in
`sd-journalist-template`, shortly to become the `sd-proxy-template`, for
installing the `securedrop-proxy` package.

Appends another task to the SLS file, ensuring that `python-apt` is
installed, because evidently Salt requires that package in order to
handle the repo configuration. Wasn't a problem against Debian 9, but
the Whonix-derived VMs such as `sd-journalist` errored out until the
package was installed.