Conditionally installs python-futures in Whonix templates
The Whonix 14 templates lack `python-futures`, due to their dependency
on an older version of the debian-9-minimal Template, so Salt cannot be
used against them out of of the box. We must ensure the package is
present in order to use Salt.

Since updating apt lists can a while, especially over Tor, let's first
check whether the package is present by trying to import it, and skip if
so. Otherwise, we'll proceed with the apt list update and install
command.

The other significant change here is more explicit dependency
declaration, so that the python-futures task is guaranteed to finish
before any Salt-related tasks are run against the child VMs.