Upgrade fedora-28 in separate qubesctl command
Salt pkg module[0] is implemented differently in Debian[1] vs Fedora[2].
the dist_upgrade parameter doesn't exist in the Fedora version and
causes the upgrade to silenty fail. `pkg.upgrade` command must be pased
without the dist_updgrade parameter.

Let's add another qubesctl command to run an update exclusively on the
fedora-28 template, in order to not have to manage update logic. We can revisit this later
if this proves unreliable or cumbersome.

[0] https://docs.saltstack.com/en/latest/ref/modules/all/salt.modules.pkg.html
[1] https://docs.saltstack.com/en/latest/ref/modules/all/salt.modules.aptpkg.html#module-salt.modules.aptpkg
[2] https://docs.saltstack.com/en/latest/ref/modules/all/salt.modules.yumpkg.html#module-salt.modules.yumpkg