Converts config tests to Python 3
We've already converted the application code to Python 3, but the config
tests (run in dom0) were still using Python 2. Fixed.

In the process, realized that the "integration" tests (targeting
`sd-journalist`) were broken. We don't run those regularly, but updated
them anyway to maximize testing. Some of that logic may be useful even
after the port to `sd-proxy`; at the very least, it's a useful test
baseline for how the submission workflow currently works.

Prunes all the colorized output because it doesn't work in dom0 (unless
using unsafe options).