Manage qubes-rpc policies and deny by default
Both templates and AppVMs provisioned by the SecureDrop workstation now
have the `sd-workstation` tag. This tag will be used to prepend a
default deny for all Qubes-RPC operations at the start of the install
process. The SecureDrop installer will prepend to this list any other
grant necessary.

sd-dom0-qvm-rpc will add default deny rule for all policies at the
end of the install to ensure that it overrides any policies that may
have been added by error as part of the other tasks.

@conorsch recommended use blockreplace for the following
reasons:
"We want to ensure that the newest config (which will change over time)
is added to the top of the file. It's OK to add additional grants
elsewhere (mostly for developers), but the Workstation-specific grants
should be enforced first.

Saltstack's "blockreplace" module allows this, and convenient supports
"markers" to ensure updates over time are inserted in the correct
location. Better yet, it's now obvious where the grants are coming from:
the securedrop-workstation config."