Moves update-templates logic to Salt
Let's not rely on the Makefile to pull in critical dependencies such as
Qubes template RPMs in dom0. We'll enforce those dependencies via Salt,
providing an update mechanism over time.

The AppArmor enabling via kernelopts is required on the sd-whonix and
sd-journalist VMs, which are cloned from the whonix-gw and whonix-ws
templates respectively, but we'll set it on the whonix workstation
templates as well, since those may be used by sd-research in the near
future.

Also culls the "prep-whonix" Makefile target, since that's been migrated
to Salt, as well.