Parallelize actions in make all
The downside to use makefile target dependencies to compose "make all"
is that the Salt actions run serially, resulting in a *much* slower
provisioning run than is strictly necessary. Salt tries to parallelize
everything; accordingly, we must explictly declare dependencies in order
to ensure that order of operations is met.

In a few instances, we had "require" statements that weren't correctly
set. Updated those to the proper syntax, and added many more explicit
dependencies based on failure output from the parallelized run.