Removes GUI feedback tooling from sd-proxy
This logic was a holdover from the sd-journalist days, in which we had
custom notification for GUI feedback in dom0. No need to continue to
configure the dependencies for those scripts, since they're not used
post conversion of the sd-proxy.

Also consolidated all sd-proxy files into the sd-proxy-template, mainly
by moving the mimetypes into a system path, as has already been done
with other VMs.