More explicit dependencies for SDW VMs
Since we're running it parallel, we must carefully declare dependencies,
so that requisites are in place before tasks assuming their presence are
run.

Uses a bit of the "onchanged" pragma to trigger commands only when
updates have happened to files on disk.

The syncing appmenus in particular tasks a long time, so it's useful to
run that only when VMs have changed.

Note that we're now copying in the submission key into the homedir of
sd-gpg; we were previously re-copying it to /tmp daily as part of the
update script; the advantage now is that we can skip the import if the
key didn't change.