Add note about supported/unsupported formats
Since we're not going to have #205 land for 0.1.0alpha, we should
explicitly call out some good formats to test with.

I've also rephrased the "contact your administrator" language --
we can re-introduce that language as we build the proper end-user
documentation, but for now, pointing to the issue tracker feels
more appropriate.