Removes Debian Jessie backports repo
The Debian Stretch VM never should have included the jessie-backports
target, and its inclusion causes apt updates to fail, blocking
installation of any and all packages. Let's snip it out.

Tried a few paths toward implementation, and it was a rocky road.
The "quiet: true" declaration is included for compatibility with the
Whonix templates, which are also based on Debian, but don't use
/etc/apt/sources.list. See the salt docs [0] for reference:

  quiet -- Do not raise any exceptions. E.g. ignore the fact that the
  file that is tried to be edited does not exist and nothing really
  happened.

However, that parameter doesn't seem to be honored when used in the SLS
files called in dom0. Another option was to wrap the jessie deletion task
in a Jinja conditional, e.g.:

  {% if 0 == salt['cmd.retcode']('test -f /etc/apt/sources.list') %}

but then the corresponding `require` for that task must *also* be
wrapped in the same conditional, otherwise Salt gets confused. That
would be a bear to maintain. Finally settled on "only if" which will
cause the task to be skipped (meaning the task passes even on Whonix)
and still available for use as a requirement.

[0] https://docs.saltstack.com/en/latest/ref/states/all/salt.states.file.html#salt.states.file.line