Refers to qubes-update-gui in README
We don't need to update the `qubes-kernel-vm-support` package in the
debian-9 template any longer, as the version shipped with 4.0.1
out of the box is sufficient for compatibility with our custom kernels.
It's still a good idea to make sure all VMs are fully updated, however,
particularly because the Salt logic currently *copies* existing VMs,
meaning that any missing updates will increase the delay on provisioning,
both initially and over time, since each copy of the TemplateVM will
need updates applied to it individually.