Documents publishing RPM repo to S3
Provides developer-oriented documentation for uploading the signed RPM
packages. The nitty-gritty details of obtaining valid AWS credentials
are glossed over, since the workflow is "talk to ops team," but
otherwise, the process is rather on-the-rails:

  1. Build the RPM (already documented)
  2. Sign the RPM (already documented)
  3. Upload the signed RPM (these docs are new)

We'll likely revise these workflows in the near future, but for now,
we'll continue to work from the docs to build familiarity with the
fundamental actions.