Renames dom0 RPM package
The catch-all name "securedrop-workstation" can refer to many different
things, including the entirety of the workstation setup. The RPM package
containing the VM config logic for setting up the various application
components is specific to dom0 (in Qubes parlance), so we'll include
dom0 in the package name, to be consistent with e.g.
`qubes-gpg-split-dom0`.