Supports dom0 RPM in dev env
Updates the Makefile targets to enable testing of the manually installed
dom0 RPM package for salt configs. Previously, the `make all` action
would clobber salt files with the contents of the locally cloned git
dir. Now, the `prep-salt` action is a bit more conservative: it will
only copy in the secrets (which are not handled by the RPM), and ignore
the salt dirs if the target dir already exist.

If `make clean` is run before `make all`, then local files will be
copied over.

Also updated the `make clean` logic to purge scripts and uninstall the
dom0 config RPM if it's present.