Enables top files in dom0 RPM postinst
By default, new Saltstack .top files (the config files that map tasks to
target hosts/VMs) will be skipped. We must explicitly enable them for
the salt runs via qubesctl to include the logic shipped in the package.
Let's port the find/enable logic from the Makefile and place that in the
RPM spec, under the `%post` macro, which is the equivalent of a `postinst`
script in a Debian package.