Installs securedrop-workstation packaging via dom0
The RPM package `qubes-template-securedrop-workstation` installs the
TemplateVM called `securedrop-workstation`. Let's update the clone
operations in the VM config to use the RPM-managed template, which will
enforce a clean image for all SDW-specific roles, as well as save a bit
of time during clean provisioning, given that the kernel packages are
already baked into the TemplateVM image, and needn't be re-fetched
during install.

There's a performance consequence to developers with this change:
first, the large (~600MB) RPM must be fetched at least once. Thereafter,
all apt packages must be updated inside the VM prior to its being cloned
for the other VMs. Will handle that in a separate commit for clarity.

Adds a few configuration tests to confirm that the RPM repo settings are
set up sanely. Note that the sys-firewall RPM repo config still doesn't
persist across reboots; we'll need to handle that separately.