Updates all apt packages inside TemplateVMs
Since we're cloning from the `qubes-template-securedrop-workstation`
image, which can be weeks or months old, we must ensure that packages
are updated before cloning it to make the various TemplateVMs, otherwise
we'll need to download the newer packages n times, where n is the number
of TemplateVMs cloned.

Mostly relevant to a clean install. Over the life of a prod instance,
updates will be managed via cron.

The update logic runs first against the `securedrop-workstation`
TemplateVM, ensuring that descendents are already updated, at least
during a clean install. The ordering is not implemented in Salt run
order: we invoke qubesctl first against just the
`securedrop-workstation` config, to force the run order we want,
rather than relying on Salt to order them appropriately.