Standardizes error handling in send-to-usb
Refactor of the script, with no changes to functionality. Motivation to
refactor is two-fold: readability and ensuring we catch all possible
exceptions. Highlights:

  * Added `exit_gracefully` def to log errors and exit 0
  * More explicit var names for constants/globals
  * Breaks logic into multiple discrete functions
  * Bottles up logic in main block, with enclosing try/except

These changes should help to ensure that we're catching as many
different exceptions as possible. The script is now slightlier
friendlier to new devs, given the slicing into functions.