Adds test for DispVM class of sd-export-usb
The `sd-export-usb` VM is a proper named DispVM, the only one in the
config to date, so let's add an assertion to the config tests to ensure
we don't lose this highly valuable property.

I've not added tests asserting that a USB device is attached because a
developer likely will not have a USB device configured and physically
plugged in at all times.