Adds config test to check for running services
Used to determine whether paxctld is running inside the VMs. The most
important is sd-svs-disp, since that's where we'll be installing
libreoffice, and libreoffice requires pax flags to be set on java dependencies
in order for the package to install successfully.

Removes the previous implementation of a service check in favor of a
more generalized approach. Overall, the config test suite should
probably be reorganized.