Ensures paxctld is running inside SDW VMs
Using Salt to force the service to run, so it's active immediately.
More important for the child VMs is using the qvm-service calls to
enable the service on boot of other, non-Template but Template-based
VMs.

Since we're now properly enforcing the paxctld service state in all VMs,
removes the previous implementation specific to sd-svs-disp.