Specify run order for paxctld and don't install recommends for Libreoffice
Setting the paxctld in the requires block of the libreoffice install will ensure the service is always started prior to installing the package.
ca-certificates-java was posing some issues at install time. As it is listed in "recommends" and not "requires", we can sidestep the issue and also reduce total amount of packages installed.