Uses dist-upgrade for apt packages
A simple "uptodate" check is insufficient to upgrade kernel versions on
the SDW VMs, because the metapackage would pull in additional packages
for the linux-{image,headers} package, an operation that requires
dist-upgrade privileges.

Also updates the apt repo config file on disk to match what's placed
inside the `qubes-template-securedrop-workstation` TemplateVM at build
time, and ships with the RPM package. (Otherwise, we'd have two apt
source files, which may cause conflicts in the future.)