Disables Whonix 14 repos
We haven't migrated to Whonix 15 yet, because we're waiting for Debian
10 to become stable in the upstream Qubes repos, so we can build and
serve all SDW-related packages via the buster channel.

Whonix has already EOL'd 14, though, so trying to run an apt update with
the old Whonix channels in place will cause apt to fail, breaking all
provisioning of the sd-whonix and sd-proxy VMs.

Let's disable those repos for now, to allow devs to continue to work,
but the priority remains landing full Buster support across the board.