References FPF Authority key in template build
We'd previously been tracking a personal staff pubkey (@emkll) for the
Qubes template build process. We've since transitioned to using the FPF
Authority Key to sign tags in the qubes-template-securedrop-workstation
repository, so the builder logic in this repo must be updated
accordingly. Done.

Used this patch during review of https://github.com/freedomofpress/qubes-template-securedrop-workstation/pull/9