Sets SD SVS as default DispVM
Previously, we weren't explicitly declaring the default DispVM template
at all. On fresh installs of 4.0.1, it'll be fedora-29, but over time,
as Fedora versions lapse into EOL, users/admins must manually update the
setting. That's no good. Rather than bump to the latest version of
Fedora (F30 as of this writing), let's use the SDW disp template, since
that includes no network, minimizing the possibility of
misconfigurations in the SDW workflows.