Applies template updates to sys-* VMs
Allows unattended-upgrades of the Fedora-based sys-{firewall,net,usb}
AppVMs. It's somewhat of a harsh approach to `qvm-kill` the VMs while
they're running, but doing so appears to be our only option, given that:

  1. It's critical these VMs stay up-to-date (to receive patches)
  2. It's not possible to modify template settings for running VMs
  3. `qvm-shutdown` won't work on netvms with attachd hosts (hence
     `qvm-kill`)

Once the configuration is applied, killing each VM once, the new
template will be used. Then we'll have to do it again in about six
months, due to Fedora release cycle.