Revisit source list removal for whonix-14 templates
Since whonix-gw-14 is used for sd-whonix directly (without a sd-whonix-template), we must apply the changes to whonix-gw-14 template to ensure sd-whonix gets, at the very least, uptream debian changes.

In the case of whonix-ws-14, we still need to edit the apt list this early in the process for first-time installs: the python-futures package needed for salt support is installed via apt, and this could break if the apt list for whonix is not commented out.

The whonix documentation recommends we us the whonix_repository command to disable the source lists.