Updates RPM signature verification
The sig verification for dom0 RPM packages was failing, due to a regex
problem with the validation lines. Presumably since F29 -> F30, SHA512
is now used, so we should honor both SHA256 and SHA512 checksums.

We plan to decommission the S3-hosting strategy for distributing RPMs,
but since I needed to make this change in order to upload the latest
Buster template RPM, I'm submitting for inclusion into master.