More robust clean action
Ensures that the various system files configured on dom0 and
sys-firewall as part of installation are removed as part of the "make
clean" action. The file cleanups are implemented in Salt.

The VM destruction has been ported from Bash to Python (via the Qubes
Admin API). The removal logic is smarter: it only removes VMs that
have the tag "sd-workstation", and removes TemplateVMs last, to avoid
ordering problems in the execution of `qvm-remove`.