Imports upstream Qubes Whonix configs
Via the `qvm.anon-whonix` SLS file, we pull in all of the following:

  * qvm.anon-whonix
  * qvm.template-whonix-ws
  * qvm.sys-whonix
  * qvm.whonix-ws-dvm
  * qvm.sys-firewall

The value of this dependency chain is that we can concisely ensure that
all Whonix-related machines, both AppVMs and TemplateVMs, are up to
date.

There's an edge case where already-running VMs, such as `sys-whonix`
(which is autostart=True by default) may not have their "template"
config field updated, since changing that value while a TemplateVM
is running is forbidden. Still, let's leverage the upstream config
as much as possible, and we'll continue to handle the
shutdown-to-change-template settings in our own scripts, as necessary.