Adds config tests for Qubes sys-* VMs
Most of the config tests to date, reasonably, target the SDW-managed
VMs. As we attempt to ensure up-to-date components across the board, we
must also ensure that Qubes-provided VM configures are updated. These
include:

  * sys-usb
  * sys-net
  * sys-firewall
  * sys-whonix

If we check those VMs to match our expected definition of "up-to-date"
templates, then we should have rather solid coverage. It's not as
complete as testing for e.g. "any fedora-based VM should be based on the
latest fedora template". Ideally we'll get to that point, but for now,
that could report failures of components unrelated to the SDW config.