Handles sys-whonix in upgrade logic
Shuts down the sys-whonix template to ensure we can update the template
settings. The Qubes-provided Salt logic only enforces template versions
on first config of the sys-whonix VM; if it already exists, it will hang
back on an older version of Whonix. That's unacceptable for us, so we'll
add a reference to the "handle-upgrade" logic to shut it down, then via
a more aggressive Salt task, enforce the new template setting.