Clean action: better handling of default dispvm
We shouldn't assume that `fedora-30-dvm` exists. It will if a developer
has manually performed the following actions:

  * install fedora-30 template
  * set fedora-30 to the default_template
  * run `qvm.default-dispvm` Salt state

If the system has fedora-30-dvm already present, then it makes sense to
use it as the default DispVM, otherwise we'll just clear out the default
DispVM setting, leaving it blank.

Reorders VM destroy step in clean-all logic

Caught by @emkll during review: it's critical that we update the default
disp_vm settings *before* we attempt to remove the VMs. Updated the
Makefile dependency ordering to accommodate.