Add security notification script sdw-notify
This script checks whether updates have been applied recently enough
using the preflight updater (AKA launcher) that is part of the
SecureDrop Workstation. "Recently enough" is defined as follows:

- If the updater has never successfully run, uptime does not exceed a
  warning threshold.
- If the updater has run successfully, it is not longer ago than the
  warning threshold. However, we allow for a grace period.

In this commit, the warning threshold is set to 5 days, and the grace
period to 30 minutes.

The script runs as a cron job and displays a PyQt warning if warranted
according to the above criteria. Lock files stored in /run/user are
used to avoid showing the warning multiple times, or while the
preflight updater is running.

As part of this commit, code from sdw-launcher has been reorganized,
to allow for shared use in a utility library by other system
components like sdw-notify.

This commit represents a collaboration between @pierwill and @eloquence.