Fixes #348 resizes private volume sises of sd-svs sd-log
The config.json now must have sizes defined for sd-svs and sd-log
private volumes.

  "vmsizes": {
    "sd_svs": 10,
    "sd_log": 5
  },

The values must be an integer.

The Makefile prep-salt target will also run the validate-config check
to make sure that size checks are done in case of:
	make sd-app
	make sd-log

The validate-config will fail in case the given size is smaller than
the current size of the private volumes in the existsing VMs.