Bugfix: Require a reboot if previous reboot was not performed
If a user were to close/exit the launcher and re-run it, they would be able to circumvent the required reboot. To ensure reboots were applied, we will now track the timestamp for the dom0 update status flag, and compare with the current system boot time. If the boot time is before the reboot flag, the launcher will ask the user to reboot, else it will apply whatever status is required.