Change JSON key "last_updated" in `sdw-update-flag` to "last_status_update"
The use of `last_updated` here is ambiguous with with the usage in the
separate file `sdw-last-udpated`, which contains the time of the
last successful update of the SecureDrop Workstation.

The `last_updated` key in `sdw-last-updated` contains the timestamp of
the last time the status itself was written to disk, regardless of
success or failure.