Use 'dev' as default environment for tests
The configure-environment script invoked by the Makefile dumps the config to /srv/salt/sd/config.json but does not update the config.json in the source folder in developement environments.

Since tests don't currently work against production or starting targets, and are run only in the source repository, let's default to dev values for tests and revisit later.