Consolidates log aggregation into single state file
We were adding multiple tasks, customized per-VM, to several state
files. Let's instead intelligently handle all logging-related config
in a single state file, and include that in the various VM
configurations as necessary.

Requires a small change to the "securedrop-log" package (managed
separately), to infer the hostname from the logclient, rather than
hardcoding it in a config file on each VM. That way, we can reuse the
same config file for all machines.

There's still an exception