Removes duplicate sd-log RPC policy
We had the same task writing twice in dom0, so removing the duplicate as
unnecessary. Also updating the tests to ensure the right policy is
configured. In updating the config tests, discovered that we didn't have
any checks for the custom (i.e. `securedrop.*`) RPC policies, so added
those. Had to adjust the testing path assumptions somewhat in order to
accommodate.