Expands config tests for logging
Ensures services are running in 'sd-log', since log aggregation can't
work if they're not. These checks are added to make the overall test
suite more durable.

In addition, updates some of the test logic to stop expecting a
hardcoded VM name in the config file. This way, we can reuse the same
config logic on all VMs.