Reboots logging VMs during install
A side-effect of consolidating the log collection logic is that all VMs,
even the TemplateVM for sd-log, begins shipping logs immediately upon
configuration. That's a problem for sd-log AppVM & TemplateVM because
the RPC calls to send logs will cause the AppVM to boot. If the AppVM
doesn't have the final config from the TemplateVM, because of an early
boot, log aggregation won't work. Let's configure both the TemplateVM &
AppVM, then shut down the TemplateVM (automatic as part of salt disp
mgmt vm logic), then reboot the AppVM. All tests passing on a clean
install.