Config test sd-whonix: verify tor config
If the torrc end state is broken on sd-whonix, we're sure to encounter it
in various failures, but having an explicit test is well worth it.
It's especially valuable to ask tor to perform the verification, given
the include-heavy config used by Whonix: there's no single torrc file to
inspect and reason about, but rather a chain of includes. Let's ask tor
to assemble and assert it's valid.