Logs stderr in VM start/stop updater actions
In an attempt to understand the specific failures during the various
actions the GUI updater runs against VMs:

  * qvm-start
  * qvm-shutdown
  * qvm-kill

Let's make sure to capture the stderr of the CLI action. We should
switch to using the Python API for most of these actions, at which time
we'll have to update the exception handling, but for now this'll give us
more visibility into the specific failures in the logs.