Logs running VMs before qvm-start
In the graphical updater, let's log the state of all running VMs
immediately prior to running `qvm-start`. That will help us
evaluate we should alter the reboot logic to avoid problems.

Really, this should be a "debug" line, not an "info" line, but since the
default logging is "info" I've left it as-is.

The original request was that we log the output of `xl list`, which has
more informative output, but it's inscrutable in log lines so I'm taking
a simpler approach that should prove as useful.