Modify update_version.sh to support rc version strings
Preserving the SecureDrop Core scheme, this will allow us to support release canadidates for the workstation provisioning code.

Pre-release separators for RPMs[1] are `~`, while not supported in Python (per PEP440)[2], and overwriting them to `-`. We should modify the RPM spec to specify the source folder and tarballs to take into account this path change.

Thanks to @kushaldas for recommending the use of `%prep` and `%setup -n <path>` here.

[1]: https://docs.fedoraproject.org/en-US/packaging-guidelines/Versioning/#_versioning_prereleases_with_tilde
[2]: https://www.python.org/dev/peps/pep-0440/#pre-release-separators