Provide release candidate verifications for update_version.sh
Because we cannot use the `~` character for prereleases due to qubes-dom0-update tooling, and we cannot use the `-` delimeter, let's use the `.` to increment version from the previous release to signal a release candidate for the next release. In the current case, the latest release is 0.2.4, the next release is 0.3.0, let's use 0.2.4.next.rc1 to signal the first release candidate for 0.3.0.