Converts F30 -> F31 in dev env
These changes ensure that fedora-31 is preferred everywhere. Running a
`make dev` will ensure that the handle-upgrade logic is run, restarting
the VMs as required. What's *not* included in this change is a method
for handling automated upgrades in staging/prod. In order to accomplish
unattended transition from f30 -> f31, we'll have to update the dom0
state management logic to "include" the handle-upgrade script.