Update Fedora TemplateVM if it has just been installed via package manager
In the Preflight updater, bacause the TemplateVM updates occur before dom0 state is applied, the Fedora-31 TemplateVM is not updated as it is installed. It will require another update cycle (8+h in order to get updated again). Here, we should simply ensure the default Fedora TemplateVM is updated, if and only if the TemplateVM was just installed.