Ensures secrets are in salt dir even in dev
The salt states require secrets such as the config file and privkey to
be stored in `/srv/salt/sd/`, as the import lines in the state files
assume those locations. The sdw-admin install action copies those files
into place, but in the case of "make clean", the secrets won't be there.
Let's make sure they are, and make sure to remove them (as on main) via
the uninstall action.