Uninstalls RPM in prep-dev script (#613)
* Uninstalls RPM in prep-dev script

Follow up to #587.

The `dnf install -y <local_file>` action will not reinstall a package
if the versions are the same. Since we expect package contents to
change, but not version strings, when running `make clone`, let's make
sure to uninstall the rpm package in dom0 so that the install action
always takes effect.