Consolidates templates via handle-upgrade script
The "securedrop-handle-upgrade" script was added in order to migrate
from Stretch-based to Buster-based templates. With some small changes to
the expected template pattern, we can leverage the same script to handle
consolidating the templates.

This only handles *creation* of the VMs, it doesn't sort out the apt
packages and config logic specific to each Template/App combination.
Additional logic will need to be added to the updater to support
hands-off consolidation.